Prototype fix for erroneous 'postcode not found' in geocoder.py working
New fallback function defined in geocode.py and called should the MoJ
postcode service not return valid data for the input postcode. Test in
place but not run as mock (and poss other dependencies) need to be added
to docker image. New class added to geocoder.py to stand in for the
postcode object returned by the MoJ postcode service API.