Models.py and migration changed for consistency with client postcode object
The Postcode class provided by the postcodeinfo API client has three
attributes which are checked when it is returned by the geocode function
in geocoder.py, 'postcode', 'longitude' and 'latitude'. Initial definition
of temporary postcodes table in Models.py used 'postcode_index' instead of
'postcode' for consistency with the tables used in the postcodeinfo service.
Following this change the object returned by objects.get() can be used
directly without copying the attributes to an instance of PostcodePlaceholder.