Reduce celery worker concurrency to 2 threads
As these workers each make a request to postcodeinfo service, 10 threads
tend to overwhelm the service.

Reducing the concurrency count seems to be a good stopgap solution.

Once we switch providers we can probably switch this back. Or at least
understand if we can do bulk latitude/longitude lookups without so many
requests.