Use Redis as the broker for Celery
We are having reproducible problems with connection health to our
currently configured CloudAMQP instances.

In short, in a few minutes the connections to the RabbitMQ broker seem
to die. If we force close them in the broker, the application thinks
everything is fine.

Vice versa, if we kill the pods, the connections stay "active" in the
broker.

We tried a Redis instance and had success. This change makes these
changes permanent.