Upgrade already installed packages
Upgrading packages as a first step would not work if our build system
used Docker layer caching. On CI, we do not use any at the moment, so
this will achieve the desired outcome.

If we ever have docker caching enabled, we will need to investigate how
to be able to upgrade packages without bloating image sizes.